refactor(dal): move WithID/DataWithID into dal as RecordWithID/RecordWithDataAndID
Relocate the typed record-id value types from package record into package dal,
renamed for clarity and to dissolve the recurring dal->record import-cycle tax
(now the typed Collection layer can return them directly in future work):

- record.WithID[K]        -> dal.RecordWithID[K]
- record.DataWithID[K,D]  -> dal.RecordWithDataAndID[K,D]

The types only ever depended on dal (Key/Record/NewRecordWithData), so moving
them adds no new import and the dal-does-not-import-record invariant still holds.

Backward compatibility: record.WithID / record.DataWithID remain as type aliases,
and record.NewWithID / record.NewDataWithID remain as thin forwarders, so
recordops and external consumers (e.g. bots-fw) compile unchanged. The embedded
field rename (DataWithID.WithID -> RecordWithDataAndID.RecordWithID) affects only
the constructor/tests being moved; no other code uses the embedded field name.

Full validation matrix ported to dal tests; record keeps forwarder tests.
golangci-lint clean; 100% coverage on moved code + forwarders; spec lint clean.
