+++ record/data_with_id.go
@@ -1,48 +1,18 @@
 package record
 
 import (
-	"fmt"
-	"reflect"
-
 	"github.com/dal-go/dalgo/dal"
 )
 
-type DataWithID[K comparable, D any] struct {
-	WithID[K]
-	Data D // we can't use *D here as consumer might want to pass an interface value instead of a pointer
-}
+// DataWithID is a backward-compatible alias for dal.RecordWithDataAndID.
+//
+// The type now lives in package dal so the typed Collection layer can return it
+// without a dal -> record import cycle. New code should prefer
+// dal.RecordWithDataAndID.
+type DataWithID[K comparable, D any] = dal.RecordWithDataAndID[K, D]
 
+// NewDataWithID forwards to dal.NewRecordWithDataAndID, kept for backward
+// compatibility.
 func NewDataWithID[K comparable, D any](id K, key *dal.Key, data D) DataWithID[K, D] {
-	if key == nil {
-		panic(fmt.Sprintf("key is nil for (id=%v)", id))
-	}
-	//if data == nil {
-	//	panic(fmt.Sprintf("data is nil for (id=%v, key=%v)", id, key))
-	//}
-	v := reflect.ValueOf(data)
-	kind := v.Kind()
-	switch kind {
-	case reflect.Pointer, reflect.Interface:
-		if v.IsNil() {
-			t := reflect.TypeOf(data)
-			panic(fmt.Sprintf("data of type %v is nil for (id=%v, key=%v)", t.String(), id, key))
-		}
-		elemType := v.Elem().Type()
-		switch elemType.Kind() {
-		case reflect.Struct, reflect.Map:
-			// OK - expected types
-		default:
-			panic("data should be a pointer to a struct or map, got " + elemType.String())
-		}
-	default:
-		t := reflect.TypeOf(data)
-		if t == nil {
-			panic(fmt.Sprintf("data is nil for (id=%v, key=%v)", id, key))
-		}
-		panic(fmt.Sprintf("data should be a pointer or an interface, got %v for (id=%v, key=%v)", t.String(), id, key))
-	}
-	return DataWithID[K, D]{
-		WithID: NewWithID(id, key, data),
-		Data:   data,
-	}
+	return dal.NewRecordWithDataAndID(id, key, data)
 }
